fix: fix arrays used as indices must be of integer

+++ fedot/core/operations/evaluation/operation_implementations/data_operations/categorical_encoders.py
@@ -99,11 +99,11 @@ class OneHotEncodingImplementation(DataOperationImplementation):
         if isinstance(features, np.ndarray):
             transformed_categorical = self.encoder.transform(features[:, self.categorical_ids]).toarray()
             # Stack transformed categorical and non-categorical data, ignore if none
-            non_categorical_features = features[:, self.non_categorical_ids]
+            non_categorical_features = features[:, self.non_categorical_ids.astype(int)]
 
         else:
             transformed_categorical = self.encoder.transform(features.iloc[:, self.categorical_ids]).toarray()
-            non_categorical_features = features.iloc[:, self.non_categorical_ids].to_numpy()
+            non_categorical_features = features.iloc[:, self.non_categorical_ids.astype(int)].to_numpy()
 
         frames = (non_categorical_features, transformed_categorical)
         transformed_features = np.hstack(frames)
